ovmerge: Support for experimental constant assignment
Constant assignment allows the use of a parameter to patch a series
of locations with fixed values, regardless of the value assigned to
the parameter. The new syntax element is an '=' at the end of a
regular parameter declaration, either followed by an in-band value
in the string or by the end of the string and a value cell (usually
a label reference).

Examples:
  // Change the target of the first fragment to be i2c1
  i2c1 = <&frag0>,"target:0=",<&i2c1>;

  // Set the compatible string for an sx1502
  sx1502 = <&sx150x>,"compatible=semtech,sx1502q";

  // Change the reg property (and the node address) to 0x21 (33)
  addr21  = <&sx150x>, "reg:0=0x21";

Signed-off-by: Phil Elwell <phil@raspberrypi.org>