ovmerge: Fix property deletion bug
The call to splice in delete_prop had only two parameters - the array
and the start position - which caused it to delete from that position
to the end, when what was required was just to delete a single value.

Fix by including the third parameter - the count - with the value
'1'.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>