ovmerge: Don't confuse zero and undefined
In Perl, the logical operator "||" returns the value of the first
operator if non-zero. Perl's concept of zero includes empty strings
and the magical NULL/None value "undef". It can be tempting to use
"||" in place of the C-like ternary operator ("?:"), but this can
go wrong if non-zero is equated with being defined as a way to provide
a default value and the defined value is actually zero (or the string
"0", since Perl automatically converts strings to integers and
vice-versa).

Signed-off-by: Phil Elwell <phil@raspberrypi.com>