ovmerge: Support literal assignments of path strings
The dtc compiler expands &node1 (as opposed to <&node1>) to the full
path to the node with the label "node1". Note that if this label is on
a node in an overlay that the path will reflect that, even if the
node with the label eventually ends up in the base dtb.

If a parameter of a base dts is used to make an alias point to another
node in the base dts, it is nice to be able to specify that replacement
path using a label on that node, rather than have to write it out
longhand. Doing so requires one simple change - treating literal
assignments of empty strings as a special case:

    console_uart0 = <&aliases>, "console=", &uart0;

Prior to this commit, this would have resulted in an error when the
dts file was parsed because the &uart0 path string would have been
interpreted as a malformed refence to the next override target.
With this commit, the parser sees the empty string assigment and
treats the next thing as a string to assign (rather like assigning
a phandle value to an integer property).

A consequence of this change is that intentionally assigning an empty
string (probably a very rare occurence) requires an explicit empty
string to be added, i.e. "...=","" or that the empty assignment falls
at the end of the list of overrides for that parameter.