vclog: Add DMA support (needed for Pi 3)
Pi 3 (BCM2710) has a 16MB hole in the ARM's address space where it can't
see RAM because the I/O space is mapped there. This is the area where
the firmware usually puts its logging information, so accessing it
from the ARM requires the use of DMA to copy it to visible RAM.

This major update adds DMA support, but also makes vclog more resilient
when run in "follow" mode.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>