dtmerge/dtoverlay: Pass all regression tests
Using dtoverlay to create a runtime overlay and merging that with a
base dtb, then comparing against merging the overlay directly with
the base dtb, showed some difference (and therefore errors).

Fix those errors by passing the stringlist length around, and by
correctly remembering the offset of the used value in a lookup table.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>