pinctrl: Extend bcm2711 up to GPIO57
BCM2711 actually has 58 GPIOs, but in most cases the additional pins are
used for RGMII and therefore not accessible as GPIOs (RGMII isn't an ALT
function, there's an override bit). CM4S is the exception to the rule,
the RGMII interface not being used, so for that reason declare all 58
GPIOs.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>