overlaycheck: Support common documentation
Add support for common groups of parameters. The name of the group is
used as the name of a pseudo-overlay where the parameters are
documented, and also as the name of a pseudo-parameter indicating
that those parameters are supported by an overlay.

For example, several updated I2C overlays are documented as having an
i2c-bus parameter, but the user is directed to read the section for the
pseudo-overlay called i2c-bus.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>