eeptools: Add the POWER_SUPPLY atom
The POWER_SUPPLY atom describes the ability of the HAT to supply
power to the Pi. This is declared in the EEPROM settings text file
using the current_supply property. The default value for current_supply
is zero, and unless a non-zero value is provided then the POWER_SUPPLY
atom is omitted from the EEPROM image.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>