dtoverlay: Adjust fixups if a phandle is deleted
Some overlays use boolean parameters as a way to delete a property. If
that property contained a phandle value (i.e. if the source had the
reference to a label on a node) then there will be a fixup that must be
deleted, otherwise the kernel will fail to apply the overlay.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>