ovmerge: Add "no renumber" option
It can be useful to pre-process and sanitise an overlay without
renumbering the fragments. It's a feature that overlaycheck uses in
order that any error messages refer to the same fragment numbers
that appear in the source files.

Add the option to prevent renumbering using "-N". N.B. There are
situations, such as combining multiple overlays, where not renumbering
may lead to non-functional output. Use with caution.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>