pinctrl: Create and use gpiolib library
Create the gpiolib library to contain all of the gpiolib code and the
gpiochips, so that applications such as pinctrl are just linked against
it, and not all of the individual object files. The named section
approach to collecting the gpiochips does not work in this situation, so
add an explicit array of pointers to the gpiochip implementations.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>