dtoverlay: Make remapping stop at the first rename
Now that it's possible to add parameters when renamei an overlay to
another, explitly stop the remap after the first rename, otherwise
it would be necessary to handle the parameters from/across multiple
renames.

Note that is possible to work around this limitation by always
mapping directly to the final name, but perhaps at the cost of
more entries in the map.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>