rpifwcrypto: Add genkey command
Add support for the new genkey firmware API which can generates
an ECDSA P-256 private key and writes it to the OTP slot for the
device unique private key.

It is intended that this should be the primary method for
generating the device unique private key. A future change may
extend the flags to create and automatically lock this key.