rpi-fw-crypto: Optimism re. HMAC input file size
Instead of determining file size for HMAC input in advance, optimistically read
up to (and including) RPI_FW_CRYPTO_HMAC_MSG_MAX_SIZE bytes with the
expectation that a short read will be performed and that the end-of-file
indicator will have been set. This allows rpi-fw-crypto to be run directly
against sysfs attributes which falsely report being page sized.